Finish tied class cohesion detection

+++ src/main/groovy/com/gitlab/artismarti/smartsmells/godclass/GodClassDetector.groovy
@@ -5,6 +5,7 @@ import com.gitlab.artismarti.smartsmells.common.Detector
 import com.gitlab.artismarti.smartsmells.common.Visitor
 
 import java.nio.file.Path
+
 /**
  * @author artur
  */
@@ -12,12 +13,13 @@ class GodClassDetector extends Detector<GodClass> {
 
 	int wmcThreshold
 	int atfdThreshold
-	BigDecimal tccThreshold
+	double tccThreshold
 
 	GodClassDetector(
 			int wmcThreshold = Defaults.WEIGHTED_METHOD_COUNT,
 			int atfdThreshold = Defaults.ACCESS_TO_FOREIGN_DATA,
-			BigDecimal tccThreshold = Defaults.TIED_CLASS_COHESION) {
+			double tccThreshold = Defaults.TIED_CLASS_COHESION) {
+
 		this.wmcThreshold = wmcThreshold
 		this.atfdThreshold = atfdThreshold
 		this.tccThreshold = tccThreshold
